safari-600.7-branch/LayoutTests/media/W3C/audio/preload/preload_reflects_metadata.html

<!doctype html>
<html>
 <head>
  <title>audio.preload - reflection test</title>
  <script src="../../w3cwrapper.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-preload">spec reference</a></p>
  <audio id="audio" preload="metadata">
  </audio>
  <div id="log"></div>
  <script>
test(function() {
 assert_equals(
  document.getElementById("audio").preload,
  "metadata",
  "audioElement.preload reflects 'metadata' value");
});
  </script>
 </body>
</html>
