safari-600.7-branch/LayoutTests/media/W3C/audio/preload/preload_reflects_metadata-expected.txt

audio.preload - reflection test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "metadata" is "metadata"

TEST COMPLETE
spec reference
